requirements.txt
asgiref==3.8.1
coverage==7.5.3
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==25.8.0
flake8==7.1.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-flake8==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
